Move proof_of_possession into bls crate
