+++ beacon_chain/utils/bls/Cargo.toml
@@ -5,3 +5,4 @@ authors = ["Paul Hauner <paul@paulhauner.com>"]
 
 [dependencies]
 bls-aggregates = { git = "https://github.com/sigp/signature-schemes" }
+hashing = { path = "../hashing" }
