+++ beacon_chain/validator_induction/src/lib.rs
@@ -3,13 +3,8 @@ extern crate hashing;
 extern crate types;
 
 mod inductor;
-mod proof_of_possession;
 
 pub use inductor::{
     ValidatorInductor,
     ValidatorInductionError,
-};
-
-pub use proof_of_possession::{
-    create_proof_of_possession,
 };
