--- beacon_chain/validator_induction/src/proof_of_possession.rs
@@ -1,22 +1,0 @@
-use bls::{
-    Keypair,
-    PublicKey,
-    Signature,
-};
-use hashing::proof_of_possession_hash;
-
-/// For some signature and public key, ensure that the signature message was the public key and it
-/// was signed by the secret key that corresponds to that public key.
-pub fn verify_proof_of_possession(sig: &Signature, pubkey: &PublicKey)
-    -> bool
-{
-    let hash = proof_of_possession_hash(&pubkey.as_bytes());
-    sig.verify_hashed(&hash, &pubkey)
-}
-
-pub fn create_proof_of_possession(keypair: &Keypair)
-    -> Signature
-{
-    let hash = proof_of_possession_hash(&keypair.pk.as_bytes());
-    Signature::new_hashed(&hash, &keypair.sk)
-}
